// packages/dash-components/src/types.d.ts
import type { EventEmitter } from '@stencil/core';

export type Scale = 's' | 'm' | 'l';
export type ScaleExtended = 's' | 'm' | 'l' | 'xl';
export type Color = 'red' | 'orange' | 'yellow' | 'green-apple' | 'green-grass' | 'baby-blue' | 'dark-blue' | 'purple' | 'pink';
export type Status = 'success' | 'error';
export type Theme = 'light' | 'dark';
export type Appearance = 'clear' | 'outline';